+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Drop `URILink` and simplify link render configuration (#11)
+- Get rid of `PandocRaw` complexity (pass raw node render as a function)
 
 ## 0.6.0.0